requirements.txt
selenium
webdriver-manager
python-dotenv
google-generativeai
